Uart update - RollingGoal

+++ Software/MotorControllerPSoC/MotorController/MotorController.cydsn/Sensor.h
@@ -11,6 +11,6 @@
 */
 #include <project.h>
 
-const uint32 * S_RPM_ptr();
+const int16 * S_RPM_ptr();
 void S_init();
 /* [] END OF FILE */
